Refactor: separate validator package
